edit quest now fills values of specific quest id in the edit form and submits edit and rerenders quests

+++ src/components/quests/QuestsList.js
@@ -1,11 +1,14 @@
 import React, { useEffect, useState } from "react";
 import { useHistory } from "react-router-dom/cjs/react-router-dom.min";
+
 // import "./PostFeed.css";
 
 export const QuestList = () => {
   const [quests, setQuests] = useState([]);
   const history = useHistory()
   const deleteQuest = (id) => {
+    if (window.confirm("Are you sure you want to delete this quest?")==false)
+      return;
     fetch(`http://localhost:8000/quests/${id}`, {
       method: "DELETE",
       headers: {
@@ -31,6 +34,7 @@ export const QuestList = () => {
 
   return (
     <>
+    <div className="mainContainer">
       <div className="headerContainer">
         <h1>Quests</h1>
       </div>
@@ -56,9 +60,14 @@ export const QuestList = () => {
                 {questObject.pillars.map(pillar => {
                   return <h4>{pillar.label}</h4>
                 })}
+                <button onClick={() => {
+                  deleteQuest(questObject.id)
+                }}>DELETE QUEST</button>
               </div>
             );
           })}
+      </div>
+      
       </div>
     </>
   );
